refactor: extract shared helpers, rename NON_INPUT_PROP_TYPES
- Extract fillDefaultsForMissingOptionalProps to mcp-utils
  (was duplicated in ap-add-step and ap-update-step)
- Extract validateAuth to mcp-utils
  (was triplicated in add/update step/trigger)
- Rename HIDDEN_PROP_TYPES to NON_INPUT_PROP_TYPES
- Remove unused imports
